allow wake queue to grow
wake queue registration is now a limited-effort workflow, which
means the registrations may grow much faster than the engine can
drain them. This puts the task runtime in the awkward position of
trying to rapid-poll insertions into the wake queue, which will
mostly lose when there are lots of subscribers.